refactor: remove crud module and update imports to repository
